LPD-84888 feat: rename inline text input component folder
