--- modules/dxp/apps/ai-hub/ai-hub-web/src/main/resources/META-INF/resources/js/content_retriever_form/inline_text_input/InlineTextInput.tsx
@@ -1,27 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import './styles.scss';
-
-import React from 'react';
-
-export default function InlineTextInput({
-	name,
-	placeholder,
-}: {
-	name: string;
-	placeholder: string;
-}) {
-	return (
-		<div className="inline-text-border">
-			<input
-				className="form-control form-control-inline"
-				name={name}
-				placeholder={placeholder}
-				type="text"
-			/>
-		</div>
-	);
-}
